Unscrambling broken position utilities
I hate Flexbox.
I hate Bootstrap's grid.
I hate the very concept of position utilities.

I just want the damn thing at the bottom of the page when you're either scrolled to the bottom or there isn't enough content to fill the page.

This should be simple, however, neither Google nor Twitter have managed to come up with an easy way to do what is probably the most common design for a website footer on the entire internet.